requirements.txt
click==8.1.2
Flask==1.1.2
pycryptodome==3.21.0
Flask_SQLAlchemy==2.4.4
pymssql==2.2.5
PyMySQL==1.0.2
requests==2.7.0
SQLAlchemy==1.3.20
scrapy==2.6.3
Jinja2==2.11.2
MarkupSafe== 1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.3
scikit-learn==1.0.2
SQLAlchemy==1.3.20
joblib==1.3.2
seaborn==0.12.2
keras==2.11.0
tensorflow==2.11.0
hdfs==2.7.0
paramiko==3.4.0
mrjob==0.7.4
